[代码改动] 1.旧版村民生成命令失效bug 2.指引hud tab冷却 3.本源方块位置 4.炼魔湾 -> 旭升湾

+++ src/main/java/fun/wraq/common/util/ClientUtils.java
@@ -593,6 +593,6 @@ public class ClientUtils {
 
     public static boolean isInBattle = false;
 
-    public static int tabSwitchLastTime = 0;
+    public static long tabSwitchLastTime = 0;
 }
